requirements.txt
anyio==4.2.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
autopep8==2.0.2
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.5.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
cryptography==42.0.2
cssselect2==0.7.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
fastjsonschema==2.19.0
fqdn==1.5.1
html5lib==1.1
idna==3.4
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
mysql-connector==2.2.9
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.2
opencv-python==4.7.0.72
oscrypto==1.3.0
overrides==7.4.0
packaging==23.2
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
Pillow==9.4.0
pip-reqs==0.11.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.17.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pypdf==4.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.12
PyYAML==6.0
pyzmq==25.1.2
qrcode==7.4.2
razorpay==1.4.1
referencing==0.32.0
reportlab==4.0.9
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.15.2
scipy==1.10.1
Send2Trash==1.8.2
six==1.16.0
sketchpy==0.1.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.3
svg.path==6.2
svglib==1.5.1
svgpathtools==1.6.0
svgwrite==1.4.3
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.64.1
traitlets==5.14.0
turtle==0.0.1
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
uritools==4.0.2
urllib3==2.0.2
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
xhtml2pdf==0.2.15
